requirements.txt
Django==1.7.1
Markdown==2.5.2
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.0.0
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
